+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "react_google_calendar",
-  "version": "2.0.1",
+  "version": "2.1.0",
   "description": "An implementation of react-big-calendar populated by linked Google calendars",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server",
-    "build": "webpack",
+    "start": "webpack-dev-server --mode development --config webpack.base.config.js --open --hot --history-api-fallback --env.PLATFORM=local --env.VERSION=stag",
+    "build": "webpack --mode production --config webpack.prod.config.js --env.PLATFORM=production --env.VERSION=stag --progress",
     "test": "jest"
   },
   "keywords": [
@@ -20,21 +20,26 @@
     "axios": "^0.16.2",
     "bluebird": "^3.5.2",
     "bootstrap": "^4.1.3",
+    "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "moment": "^2.22.2",
     "react": "^15.6.1",
     "react-big-calendar": "^0.14.4",
-    "react-dom": "^15.6.1",
-    "webpack-dev-server": "^2.11.3"
+    "react-dom": "^15.6.1"
   },
   "devDependencies": {
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.3.2",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-2": "^6.13.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/polyfill": "^7.0.0-beta.51",
+    "@babel/preset-env": "^7.0.0-beta.51",
+    "@babel/preset-react": "^7.0.0-beta.51",
+    "babel-loader": "^8.0.0-beta.0",
+    "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.11",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^23.6.0",
@@ -42,8 +47,11 @@
     "postcss-loader": "^2.1.6",
     "sass-loader": "^6.0.7",
     "style-loader": "^0.18.2",
-    "webpack": "^3.12.0",
-    "webpack-dev-middleware": "^1.12.2"
+    "webpack": "^4.12.1",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-middleware": "^1.12.2",
+    "webpack-dev-server": "^3.1.0",
+    "webpack-merge": "^4.1.4"
   },
   "repository": {
     "type": "git",
