+++ src/public/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="text/javascript" src="js/bundle.js"></script>
-  </body>
+    <script src="./public/dist/main.bundle.js"></script>
+  <script type="text/javascript" src="app.bundle.js"></script></body>
 </html>
